Genetic algorithm but user specific
This performs the genetic algorithm across all users (17 in our current data set). Each user's total app list is split into a day-by-day list, and the genetic algorithm generates random lists of apps to compare to the sequences in the user's day-by-day lists. The algorithm is self-refining and outputs JSON result files . 'results' files contain the optimal app list, 'apps' files contain the follow set for all the apps in the optimal app list for that user.